Split JS into class files
- Added 3 class files (UI, Storage & Book)
