+++ index.html
@@ -7,7 +7,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>Awesome Books</title>
   <link rel="stylesheet" href="index.css">
-  <script src="scripts/app.js" defer></script>
+  <script type="module" src="./scripts/app.js" defer></script>
 </head>
 
 <body>
