+++ scripts/app.js
@@ -1,71 +1,7 @@
-/* eslint-disable max-classes-per-file */
 /* eslint-disable no-alert */
-
-// Book Class
-class Book {
-  constructor(title, author) {
-    this.title = title;
-    this.author = author;
-  }
-}
-
-// Storage Class
-class Storage {
-  static getBooks() {
-    let books = [];
-    if (localStorage.getItem('data')) {
-      books = JSON.parse(localStorage.getItem('data'));
-    }
-
-    return books;
-  }
-
-  static addBook(b) {
-    const books = Storage.getBooks();
-    books.push(b);
-    localStorage.setItem('data', JSON.stringify(books));
-  }
-
-  static deleteBook(textContent) {
-    const books = Storage.getBooks();
-
-    books.forEach((book, index) => {
-      if (textContent.includes(book.title) && textContent.includes(book.author)) {
-        books.splice(index, 1);
-      }
-    });
-    localStorage.setItem('data', JSON.stringify(books));
-  }
-}
-
-// UI Class
-class UI {
-  static displayBooks() {
-    const books = Storage.getBooks();
-    books.forEach((book) => UI.addBookToList(book));
-  }
-
-  static addBookToList(b) {
-    const bookList = document.getElementById('bookList');
-    const row = document.createElement('tr');
-    row.innerHTML = `
-      <td>"${b.title}" by ${b.author}</td>
-      <td><button type='button' class="remove">Remove</button></td>`;
-
-    bookList.appendChild(row);
-  }
-
-  static removeBook(e) {
-    if (e.classList.contains('remove')) {
-      e.parentElement.parentElement.remove();
-    }
-  }
-
-  static clearFields() {
-    document.getElementById('bookTitle').value = '';
-    document.getElementById('author').value = '';
-  }
-}
+import {UI} from './ui.js';
+import {Book} from './book.js';
+import {Storage} from './storage.js'
 
 // Event: display books
 document.addEventListener('DOMContentLoaded', UI.displayBooks);
